+++ lib/majic/helpers.ex
@@ -1,10 +1,10 @@
-defmodule GenMagic.Helpers do
+defmodule Majic.Helpers do
   @moduledoc """
   Contains convenience functions for one-off use.
   """
 
-  alias GenMagic.Result
-  alias GenMagic.Server
+  alias Majic.Result
+  alias Majic.Server
 
   @spec perform_once(Path.t() | {:bytes, binary}, [Server.option()]) ::
           {:ok, Result.t()} | {:error, term()}
@@ -16,9 +16,9 @@ defmodule GenMagic.Helpers do
 
   ## Example
 
-      iex(1)> {:ok, result} = GenMagic.Helpers.perform_once(".")
+      iex(1)> {:ok, result} = Majic.Helpers.perform_once(".")
       iex(2)> result
-      %GenMagic.Result{content: "directory", encoding: "binary", mime_type: "inode/directory"}
+      %Majic.Result{content: "directory", encoding: "binary", mime_type: "inode/directory"}
   """
   def perform_once(path, options \\ []) do
     with {:ok, pid} <- Server.start_link(options),