+++ CHANGELOG.md
@@ -7,7 +7,24 @@ The format is based on [Keep a Changelog][1], and this project adheres to [Seman
 [1]: https://keepachangelog.com/en/1.0.0/
 [2]: https://semver.org/spec/v2.0.0.html
 
-## [Unreleased]
+## majic [Unreleased]
+
+## Added
+
+- Forked gen_magic.
+- Pool: `Majic.Pool`
+- Unified API: `Majic.perform/1,2,3`
+
+## Changed
+
+- C port now using erl_interface
+- `Majic.Server.reload/2,3`
+- `Majic.Server.recycle/2,3`
+- Bytes support: `Majic.Server.perform(ref, {:bytes, <<>>})`
+- Builds on Musl
+- Better error and timeout handling
+
+## gen_majic [1.0]
 
 ### Added